+++ apps/ssr/src/app/app.component.ts
@@ -1,26 +1,34 @@
 import { isPlatformServer } from '@angular/common';
 import { Component, Inject, NgZone, OnInit, PLATFORM_ID } from '@angular/core';
-import { BehaviorSubject } from 'rxjs';
+import { rxState } from '@rx-angular/state';
+import { of } from 'rxjs';
 
 @Component({
   selector: 'rx-angular-root',
   template: `
-    <div id="let" *rxLet="color$ as color; strategy">{{ color }}</div>
+    <div id="let" *rxLet="color$; let color">{{ color }}</div>
     <div id="push">{{ color$ | push }}</div>
     <div id="unpatch" [unpatch]="['click']" (click)="onClick()"></div>
     <div class="for" *rxFor="let color of colors$">{{ color }}</div>
   `,
 })
 export class AppComponent implements OnInit {
-  color$ = new BehaviorSubject('red');
-  colors$ = new BehaviorSubject(['red']);
+  private readonly state = rxState<{ color: string; colors: string[] }>(
+    ({ set, connect }) => {
+      set('color', () => 'red');
+      connect('colors', of(['red']));
+    }
+  );
+
+  readonly color$ = this.state.select('color');
+  readonly colors$ = this.state.select('colors');
 
   constructor(@Inject(PLATFORM_ID) private platformId: string) {}
 
   ngOnInit() {
     if (isPlatformServer(this.platformId)) {
-      this.color$.next('green');
-      this.colors$.next(['green', 'purple']);
+      this.state.set('color', () => 'green');
+      this.state.set('colors', () => ['green', 'purple']);
     }
   }