+++ libs/cdk/coalescing/src/lib/index.ts
@@ -1,3 +1,0 @@
-export { RxCoalescingOptions, coalescingObj } from './model';
-export { coalesceWith } from './coalesceWith';
-export { coalescingManager, CoalescingManager } from './coalescingManager';