+++ libs/template/if/src/lib/tests/if.directive.context.spec.ts
@@ -6,7 +6,7 @@ import {
 } from '@angular/core/testing';
 import { RxNotificationKind } from '@rx-angular/cdk/notifications';
 import { RX_RENDER_STRATEGIES_CONFIG } from '@rx-angular/cdk/render-strategies';
-import { mockConsole } from '@test-helpers';
+import { mockConsole } from '@test-helpers/rx-angular';
 import {
   BehaviorSubject,
   concat,