+++ libs/template/let/src/lib/tests/let.directive.template-binding.no-suspense.spec.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RX_RENDER_STRATEGIES_CONFIG } from '@rx-angular/cdk/render-strategies';
-import { mockConsole } from '@test-helpers';
+import { mockConsole } from '@test-helpers/rx-angular';
 import { Observable, of, Subject } from 'rxjs';
 import { RxLet } from '../let.directive';