+++ libs/test-helpers/src/index.ts
@@ -1,1 +1,3 @@
-export * from './lib/index';
+export * from './lib/fixtures/fixtures';
+export * from './lib/rx-marbles';
+export * from './lib/utils';